chore: remove setTimeout usages when app is in background
